+++ terraform/core/catalog.tf
@@ -6,6 +6,7 @@ resource "helm_release" "catalog-application" {
   namespace  = "core"
   create_namespace = false
   timeout = var.helm_timeout
+  force_update = true
 
   set {
     name  = "replicaCount"
@@ -39,7 +40,6 @@ resource "helm_release" "catalog-batch" {
   repository = var.helm_repository
   name       = "catalog-batch"
   chart      = "${var.helm_repository}/catalog/batch"
-  version    = "1.1.2"
   namespace  = "core"
   create_namespace = false
   timeout = var.helm_timeout