+++ terraform/example/provisioning/person-service.tf
@@ -1,0 +1,57 @@
+locals {
+  helm_releases = {
+    application = { provisioning_enabled = "false" }
+    provisioning = {  provisioning_enabled = "true" }
+  }
+}
+
+resource "helm_release" "person-service-application" {
+  repository         = var.helm_repository
+  for_each           = local.helm_releases
+  name               = "person-service-application-${each.key}"
+  chart            = "${var.helm_repository}/person-service/application"
+  namespace        = "example"
+  create_namespace = false
+  timeout          = var.helm_timeout
+  depends_on = [helm_release.person-service-application["provisioning"]]
+
+  set {
+    name  = "replicaCount"
+    value = "1"
+  }
+
+  set {
+    name  = "ingress.hosts"
+    value = var.hostname
+  }
+
+  set {
+    name  = "image.arch"
+    value = strcontains(var.helm_repository, "spring") ? "-native${local.server_arch}" : local.server_arch
+  }
+
+  set_sensitive {
+    name  = "database.password"
+    value = random_password.postgres_password.result
+  }
+
+  set {
+    name = "oidc.enabled"
+    value = local.oidc_enabled
+  }
+
+  set_sensitive {
+    name  = "oidc.session.secret"
+    value = random_password.oidc_session_secret.result
+  }
+
+  set {
+    name = "multiTenancy.tenants"
+    value = var.multi_tenancy_tenants
+  }
+
+  set {
+    name  = "provisioning.enabled"
+    value = each.value.provisioning_enabled
+  }
+}