+++ terraform/example/variables.tf
@@ -20,3 +20,9 @@ locals {
   oidc_enabled = strcontains(var.hostname, ".de")
   postgres_ha = false
 }
+
+# terraform taint helm_release.person-service-application
+# terraform taint 'helm_release.person-service-application["provisioning"]'
+variable "multi_tenancy_tenants" {
+  default = "0\\,5\\,8"
+}